feat: add generative objectives, samplers and few-step distillation
Replace Lightning with native PyTorch training, validation, logging and
resumable checkpoints, including EMA, RNG and auxiliary optimizer state.

Add epsilon, x0, v, score and velocity targets with VP, VE, sub-VP, linear,
cosine and EDM trajectories. Support rectified flow, flow matching,
consistency training and teacher distillation, plus extended samplers
including PNDM, PLMS, UniPC, DEIS and DPM-Solver variants.

Implement DMD, DMD2 with alternating critic updates and backward simulation,
progressive DDIM distillation, paired trajectory regression and reflow.
Enforce compatible teacher, prediction and sampling-grid contracts.

Add model presets, mathematical and integration tests, usage documentation
and updated notebooks. Correct DDIM updates for the selected timestep grid,
preserve embedding dtypes and ignore the local uv.lock file.

Validation: 160 tests pass; Ruff E/F checks, formatting and notebook syntax
checks pass. CPU training, checkpoint resume and standalone student sampling
are verified; converged sample quality is not evaluated.

+++ src/diffusion/layers/embed.py
@@ -59,7 +59,7 @@ class SinusoidalEncoding(nn.Module):
         device = t.device
         batch_size = t.shape[0]
 
-        emb = torch.zeros(batch_size, self.embed_dim, device=device)
+        emb = torch.zeros(batch_size, self.embed_dim, device=device, dtype=self.omega.dtype)
         emb[:, 0::2] = torch.sin(self.omega * t)
         emb[:, 1::2] = torch.cos(self.omega * t)
